+++ spec/ah/lane/resource.js
@@ -18,9 +18,7 @@ module.exports = function( host ) {
 					}
 				},
 				handle: function( envelope ) {
-					var lane = _.cloneDeep( _.where( model.board1.lanes, { id: parseInt( envelope.data.laneId ) } )[ 0 ] );
-					lane.boardId = model.board1.id;
-					// lane.board = { id: model.board1.id };
+					var lane = _.cloneDeep( _.where( model.board1.lanes, { id: parseInt( envelope.data.id ) } )[ 0 ] );
 					envelope
 						.hyped( lane )
 						.render();
@@ -29,9 +27,10 @@ module.exports = function( host ) {
 			cards: {
 				method: "get",
 				url: "/lane/:lane.id/card",
+				include: [ "id", "title", "wip" ],
 				handle: function( envelope ) {
 					envelope
-						.hyped( _.where( model.board1.lanes, { id: parseInt( envelope.data.laneId ) } )[ 0 ].cards )
+						.hyped( _.where( model.board1.lanes, { id: parseInt( envelope.data.id ) } )[ 0 ].cards )
 						.resource( "card" )
 						.action( "self" )
 						.render();