+++ spec/filter.spec.js
@@ -1,12 +1,8 @@
 var should = require( "should" ); // jshint ignore: line
-var _ = require( "lodash" );
-var when = require( "when" );
 var model = require( "./model.js" );
-var HyperModel = require( "../src/hyperModel.js" );
+var hyperResource = require( "../src/hyperResource.js" );
 
 var board1 = model.board1;
-var board2 = model.board2;
-var deepCompare = model.deepCompare;
 
 describe( "with property filter", function() {
 	var resource = {
@@ -35,11 +31,11 @@ describe( "with property filter", function() {
 	};
 
 	before( function() {
-		var hypermodel = HyperModel( { board: resource } );
-		self = hypermodel( board1, "board", "self" ).render();
+		var fn = hyperResource.renderFn( { board: resource } );
+		self = fn( "board", "self", board1 );
 	} );
 
-	it( 'should generate self hypermedia object model', function() {
+	it( "should generate hypermedia without `lanes` or `hidden` fields", function() {
 		self.should.eql( expectedSelf );
 	} );
 } );