+++ src/hyperResponse.js
@@ -1,8 +1,8 @@
-var HyperResponse = function( req, res, engine, hyperModel, contentType ) {
+var HyperResponse = function( req, res, engine, hyperResource, contentType ) {
 	this._authCheck = req._checkPermission || function() { return true; };
 	this._code = 200;
 	this._engine = engine;
-	this._hyperModel = hyperModel;
+	this._hyperResource = hyperResource;
 	this._req = req;
 	this._res = res;
 	this._contentType = contentType;
@@ -14,12 +14,6 @@ var HyperResponse = function( req, res, engine, hyperModel, contentType ) {
 		self._context = context;
 		return self;
 	};
-};
-
-HyperResponse.prototype.action = function( action ) {
-	this._action = action;
-	this._originUrl = undefined;
-	return this;
 };
 
 HyperResponse.prototype.context = function( context ) {
@@ -38,25 +32,24 @@ HyperResponse.prototype.status = function( code ) {
 	return this;
 };
 
-HyperResponse.prototype.resource = function( resource ) {
-	this._resource = resource;
-	this._originUrl = undefined;
-	return this;
-};
-
 HyperResponse.prototype.render = function() {
 	if( this._engine ) {
+		try {
 		var resource = this._req._resource;
 		var action = this._req._action;
-		var hypermedia = this._hyperModel( this._model, resource, action )
-			.useResource( this._resource )
-			.useAction( this._action )
-			.useOrigin( this._originUrl, this._originMethod )
-			.useContext( this._context )
-			.auth( this._authCheck )
-			.render();
+		var hypermedia = this._hyperResource( 
+			this._resource || resource,
+			this._action || action,
+			this._model,
+			"",
+			this._context,
+			this._originUrl,
+			this._originMethod
+		);
+
 		var body = this._engine( hypermedia );
 		this._res.set( "Content-Type", this._contentType ).status( this._code ).send( body );
+	} catch( e ) { console.log( e.stack ); }
 	} else {
 		this._res.status( 415 ).send( "The requested media type '" + this._contentType + "' is not supported. Please see the OPTIONS at the api root to get a list of supported types." );
 	}